Support explicit declaration of multivalued fields.
Multivalued fields, and only multivalued fields, are assumed to come out of
ElasticSearch as arrays. Otherwise the last value for the field will clobber
previous values.

This makes things more predictable with ElasticSearch >= 1.0, where the _source
of a document will be a singleton array if the document was inserted with a
singleton array.